move object related tests to their own class
