+++ Assets/Tests/Runtime/NetworkObjectManagerTest.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c0a1aaa72952198428cc6e0da9ae5d4d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
